requirements.txt
streamlit==1.29.0
boto3==1.33.6
python-dotenv==1.0.0
pillow==10.1.0
pycognito==2023.11.0
streamlit-authenticator==0.2.3